// src/multiselect.interface.ts
export interface DropdownSettings{
    singleSelection: Boolean;
    text: String;
    enableCheckAll : Boolean;
    selectAllText: String;
    unSelectAllText: String;
    enableSearchFilter: Boolean;
    maxHeight: Number;
    badgeShowLimit: Number;
    classes: String;
    limitSelection?: Number;
    disabled?: Boolean;
    searchPlaceholderText: String;
    groupBy?: String;
    showCheckbox?: Boolean;
    noDataLabel: string;
    searchAutofocus?: boolean;
    value?: string;
    description?: string;

} 